test(odoh): move the budget tests inline
They only exercise odoh.rs, so they belong in its #[cfg(test)] module
next to cache_backoff_blocks_refetch_after_failure, which covers the
same backoff. soa_compression_bug stays the one out-of-module test.

Being in-crate also lets the backoff test assert last_failure directly
instead of matching on "backoff active" in an error string.
